Don't trailingslashit() the return of participad_api_endpoint() if it's empty
The trailingslashit() call was introduced in the previous version, to ensure
that proper endpoints always contained the trailing slash, as required by the
API spec. However, this resulted in the default value (an empty string) being
slashed, with the result that it was no longer empty, and Participad tried
(and failed) to create a new client instance.

Fixes #25